requirements.txt
# Core Dependencies
numpy>=1.26.4
pandas>=2.2.0
matplotlib>=3.8.2
seaborn>=0.13.1

# Reinforcement Learning Framework
gymnasium>=0.29.1
stable-baselines3[extra]>=2.2.1
torch>=2.1.2

# Data Processing
scikit-learn>=1.3.0
scipy>=1.11.0
requests>=2.31.0
pyyaml>=6.0.1
openpyxl>=3.1.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# API & Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
python-multipart>=0.0.6

# Monitoring & Logging
prometheus-client>=0.19.0
tensorboard>=2.15.0
tqdm>=4.66.0
